usps-chatbot/requirements.txt
# Web server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Claude API
anthropic>=0.49.0

# Supabase vector store
supabase>=2.3.0

# Embeddings
voyageai>=0.2.3

# Utilities
python-dotenv>=1.0.0
numpy>=1.26.0
httpx>=0.27.0
